backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Environment variables
python-dotenv==1.0.1

# Playwright (LinkedIn scraping)
playwright==1.49.0

# Scheduler
apscheduler==3.10.4

# AI – Claude
anthropic==0.40.0

# AI – Gemini
google-generativeai==0.8.3

# Validation / data
pydantic==2.10.3

# HTTP (utility)
httpx==0.28.1
